# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sorteio Ativo",
    "version": "1.0.0",
    "description": "Documentação da API do Sorteio Ativo"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/orders": {
      "post": {
        "summary": "Finaliza uma compra e marca números como vendidos",
        "description": "Esta API é chamada após um pagamento bem-sucedido (ex: retorno do Mercado Pago). Ela verifica a disponibilidade final dos números (considerando se já estão reservados/vendidos pelo mesmo usuário ou por outros). Se tudo estiver OK, marca os números como 'sold' e atualiza o pedido para 'completed' (se ainda não estiver).\n",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId",
                  "selectedNumbers",
                  "internalOrderId"
                ],
                "properties": {
                  "productId": {
                    "type": "integer",
                    "description": "ID do produto."
                  },
                  "selectedNumbers": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "Lista dos números selecionados."
                  },
                  "internalOrderId": {
                    "type": "integer",
                    "description": "O ID do pedido que foi previamente criado com status 'pending'."
                  },
                  "paymentDetails": {
                    "type": "string",
                    "description": "Detalhes do pagamento (ex: ID do pagamento do Mercado Pago).\n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido finalizado com sucesso, números marcados como vendidos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "orderId": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida (dados faltando, etc.)."
          },
          "401": {
            "description": "Não autenticado."
          },
          "404": {
            "description": "Produto ou Pedido não encontrado."
          },
          "409": {
            "description": "Conflito (um ou mais números selecionados não estão disponíveis para este usuário).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "unavailable_numbers": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Retorna os detalhes de um produto específico e seus números",
        "description": "Busca os detalhes de um produto pelo seu ID, juntamente com a lista completa de seus números e seus respectivos status (disponível, vendido, etc.).\n",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "O ID numérico do produto.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do produto e a lista de números.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "price_per_number": {
                          "type": "number"
                        },
                        "numbers": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "number_value": {
                                "type": "integer"
                              },
                              "status": {
                                "type": "string"
                              },
                              "user_id": {
                                "type": "integer",
                                "nullable": true
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
...
  },
...
}